minor changes to infix explanation
changed `called` to `expressed` , because `call` sound more like a function call. Also mentioned the term operator.

Additionally added a sentence how this works internally. This was taught in the "Introduction to AI with Prolog" lecture. Therefore, might be easier for some students to make the connection that this is the same.